testing Hard Disk access using PIO and AT interface.

+++ mm.c
@@ -54,7 +54,7 @@ void print_heap_magic() {
 }
 
 byte * kmalloc(size_t nblks) {
-	if(free_ptr + nblks > MAX_BLOCKS) 
+	if((size_t)free_ptr + nblks > MAX_BLOCKS) 
 		return NULL;
 	byte *tmp = free_ptr;
 	free_ptr = free_ptr + nblks + 1;
